refactor: move registration api to identity-service
