+++ core_apps/identity/views.py
@@ -4,7 +4,8 @@ from rest_framework_simplejwt.views import TokenRefreshView, TokenObtainPairView
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
 from rest_framework.response import Response
-from core_apps.user import serializers
+from rest_framework.generics import CreateAPIView
+from . import serializers
 
 
 class MyTokenRefreshView(TokenRefreshView):
@@ -22,3 +23,7 @@ class ProfileFromTokenView(APIView):
         user = request.user
         serializer = serializers.ReadUpdateUserProfile(user)
         return Response(serializer.data)
+
+
+class ProfileRegistration(CreateAPIView):
+    serializer_class = serializers.RegisterUser
