+++ core_apps/user/urls.py
@@ -9,7 +9,6 @@ profile_router.register('', views.UserProfileViewset, "profile")
 
 urlpatterns = [
     path('profile/search/', views.UserProfileSearch.as_view()),
-    path('profile/registration/', views.ProfileRegistration.as_view()),
     path('profile/ids/', views.UserProfileByIds.as_view()),
     path('profile/', include(profile_router.urls)),
     path('profile/delete/', views.DeleteUserProfile.as_view()),
